[WIP] Seems to work, but the issue still happens after a reload and start some tasks
Update the generation of the terminal id when executing tasks
Currently, when tasks are executed, we use the TerminalWidgetFactoryOptions
in order to create terminals to attach to and display the output of the
task's execution. One of the available properties is the id field which
is prefixed by the string terminal-.
The problem is that when opening terminals the same prefix (terminal-)
is also used which may potentially lead to collisions. In order to
address the issue, terminals which are created for tasks are instead
prefixed by the string task-.

Signed-off-by: Jacques Bouthillier <jacques.bouthillier@ericsson.com>